fix: Properly fetch type of full None List Series
Fixes #18910, 18911.
