+++ crates/polars-core/src/frame/column/mod.rs
@@ -1,10 +1,10 @@
 use std::borrow::Cow;
-use std::sync::OnceLock;
 
 use num_traits::{Num, NumCast};
 use polars_error::PolarsResult;
 use polars_utils::index::check_bounds;
 use polars_utils::pl_str::PlSmallStr;
+pub use scalar::ScalarColumn;
 
 use self::gather::check_bounds_ca;
 use crate::chunked_array::cast::CastOptions;
@@ -15,6 +15,7 @@ use crate::utils::{slice_offsets, Container};
 use crate::{HEAD_DEFAULT_LENGTH, TAIL_DEFAULT_LENGTH};
 
 mod arithmetic;
+mod scalar;
 
 /// A column within a [`DataFrame`].
 ///
@@ -32,25 +33,6 @@ mod arithmetic;
 pub enum Column {
     Series(Series),
     Scalar(ScalarColumn),
-}
-
-/// A [`Column`] that consists of a repeated [`Scalar`]
-///
-/// This is lazily materialized into a [`Series`].
-#[derive(Debug, Clone)]
-pub struct ScalarColumn {
-    name: PlSmallStr,
-    // The value of this scalar may be incoherent when `length == 0`.
-    scalar: Scalar,
-    length: usize,
-
-    // invariants:
-    // materialized.name() == name
-    // materialized.len() == length
-    // materialized.dtype() == value.dtype
-    // materialized[i] == value, for all 0 <= i < length
-    /// A lazily materialized [`Series`] variant of this [`ScalarColumn`]
-    materialized: OnceLock<Series>,
 }
 
 /// Convert `Self` into a [`Column`]
@@ -97,7 +79,11 @@ impl Column {
         match self {
             Column::Series(s) => s,
             Column::Scalar(s) => {
-                let series = s.materialized.take().unwrap_or_else(|| s.to_series());
+                let series = std::mem::replace(
+                    s,
+                    ScalarColumn::new_empty(PlSmallStr::EMPTY, DataType::Null),
+                )
+                .take_materialized_series();
                 *self = Column::Series(series);
                 let Column::Series(s) = self else {
                     unreachable!();
@@ -121,7 +107,7 @@ impl Column {
     pub fn dtype(&self) -> &DataType {
         match self {
             Column::Series(s) => s.dtype(),
-            Column::Scalar(s) => s.scalar.dtype(),
+            Column::Scalar(s) => s.dtype(),
         }
     }
 
@@ -129,8 +115,8 @@ impl Column {
     pub fn field(&self) -> Cow<Field> {
         match self {
             Column::Series(s) => s.field(),
-            Column::Scalar(s) => match s.materialized.get() {
-                None => Cow::Owned(Field::new(s.name.clone(), s.scalar.dtype().clone())),
+            Column::Scalar(s) => match s.lazy_as_materialized_series() {
+                None => Cow::Owned(Field::new(s.name().clone(), s.dtype().clone())),
                 Some(s) => s.field(),
             },
         }
@@ -140,7 +126,7 @@ impl Column {
     pub fn name(&self) -> &PlSmallStr {
         match self {
             Column::Series(s) => s.name(),
-            Column::Scalar(s) => &s.name,
+            Column::Scalar(s) => s.name(),
         }
     }
 
@@ -148,7 +134,7 @@ impl Column {
     pub fn len(&self) -> usize {
         match self {
             Column::Series(s) => s.len(),
-            Column::Scalar(s) => s.length,
+            Column::Scalar(s) => s.len(),
         }
     }
 
@@ -162,13 +148,7 @@ impl Column {
     pub fn rename(&mut self, name: PlSmallStr) {
         match self {
             Column::Series(s) => _ = s.rename(name),
-            Column::Scalar(s) => {
-                if let Some(series) = s.materialized.get_mut() {
-                    series.rename(name.clone());
-                }
-
-                s.name = name;
-            },
+            Column::Scalar(s) => _ = s.rename(name),
         }
     }
 
@@ -298,7 +278,7 @@ impl Column {
     pub fn clear(&self) -> Self {
         match self {
             Column::Series(s) => s.clear().into(),
-            Column::Scalar(s) => Self::new_scalar(s.name.clone(), s.scalar.clone(), 0),
+            Column::Scalar(s) => s.resize(0).into(),
         }
     }
 
@@ -315,8 +295,8 @@ impl Column {
         match self {
             Column::Series(s) => s.new_from_index(index, length).into(),
             Column::Scalar(s) => {
-                if index >= s.length {
-                    Self::full_null(s.name.clone(), length, s.scalar.dtype())
+                if index >= s.len() {
+                    Self::full_null(s.name().clone(), length, s.dtype())
                 } else {
                     s.resize(length).into()
                 }
@@ -336,14 +316,18 @@ impl Column {
     pub fn is_null(&self) -> BooleanChunked {
         match self {
             Self::Series(s) => s.is_null(),
-            Self::Scalar(s) => BooleanChunked::full(s.name.clone(), s.scalar.is_null(), s.length),
+            Self::Scalar(s) => {
+                BooleanChunked::full(s.name().clone(), s.scalar().is_null(), s.len())
+            },
         }
     }
     #[inline]
     pub fn is_not_null(&self) -> BooleanChunked {
         match self {
             Self::Series(s) => s.is_not_null(),
-            Self::Scalar(s) => BooleanChunked::full(s.name.clone(), !s.scalar.is_null(), s.length),
+            Self::Scalar(s) => {
+                BooleanChunked::full(s.name().clone(), !s.scalar().is_null(), s.len())
+            },
         }
     }
 
@@ -370,7 +354,7 @@ impl Column {
         match self {
             Column::Series(s) => s.slice(offset, length).into(),
             Column::Scalar(s) => {
-                let (_, length) = slice_offsets(offset, length, s.length);
+                let (_, length) = slice_offsets(offset, length, s.len());
                 s.resize(length).into()
             },
         }
@@ -386,7 +370,7 @@ impl Column {
     pub fn null_count(&self) -> usize {
         match self {
             Self::Series(s) => s.null_count(),
-            Self::Scalar(s) if s.scalar.is_null() => s.length,
+            Self::Scalar(s) if s.scalar().is_null() => s.len(),
             Self::Scalar(_) => 0,
         }
     }
@@ -796,7 +780,7 @@ impl Column {
 
         match self {
             Column::Series(s) => s.gather_every(n, offset).into(),
-            Column::Scalar(s) => s.resize(s.length - offset / n).into(),
+            Column::Scalar(s) => s.resize(s.len() - offset / n).into(),
         }
     }
 
@@ -812,7 +796,7 @@ impl Column {
         match self {
             Column::Series(s) => s.extend_constant(value, n).map(Column::from),
             Column::Scalar(s) => {
-                if s.scalar.as_any_value() == value {
+                if s.scalar().as_any_value() == value {
                     Ok(s.resize(s.len() + n).into())
                 } else {
                     s.as_materialized_series()
@@ -879,7 +863,7 @@ impl Column {
 
         match self {
             Column::Series(s) => s.get_unchecked(index),
-            Column::Scalar(s) => s.scalar.as_any_value(),
+            Column::Scalar(s) => s.scalar().as_any_value(),
         }
     }
 
@@ -1010,237 +994,13 @@ impl PartialEq for Column {
 impl From<Series> for Column {
     #[inline]
     fn from(series: Series) -> Self {
+        // We instantiate a Scalar Column if the Series is length is 1. This makes it possible for
+        // future operations to be faster.
         if series.len() == 1 {
-            // SAFETY: We just did the bounds check
-            let value = unsafe { series.get_unchecked(0) }.into_static();
-            let value = Scalar::new(series.dtype().clone(), value);
-            let mut col = ScalarColumn::new(series.name().clone(), value, 1);
-            col.materialized = OnceLock::from(series);
-            return Self::Scalar(col);
+            return Self::Scalar(ScalarColumn::unit_scalar_from_series(series));
         }
 
         Self::Series(series)
-    }
-}
-
-impl From<ScalarColumn> for Column {
-    #[inline]
-    fn from(value: ScalarColumn) -> Self {
-        Self::Scalar(value)
-    }
-}
-
-impl ScalarColumn {
-    #[inline]
-    pub fn new(name: PlSmallStr, scalar: Scalar, length: usize) -> Self {
-        Self {
-            name,
-            scalar,
-            length,
-
-            materialized: OnceLock::new(),
-        }
-    }
-
-    #[inline]
-    pub fn new_empty(name: PlSmallStr, dtype: DataType) -> Self {
-        Self {
-            name,
-            scalar: Scalar::new(dtype, AnyValue::Null),
-            length: 0,
-
-            materialized: OnceLock::new(),
-        }
-    }
-
-    pub fn name(&self) -> &PlSmallStr {
-        &self.name
-    }
-
-    pub fn dtype(&self) -> &DataType {
-        self.scalar.dtype()
-    }
-
-    pub fn len(&self) -> usize {
-        self.length
-    }
-
-    pub fn is_empty(&self) -> bool {
-        self.length == 0
-    }
-
-    fn _to_series(name: PlSmallStr, value: Scalar, length: usize) -> Series {
-        let series = if length == 0 {
-            Series::new_empty(name, value.dtype())
-        } else {
-            value.into_series(name).new_from_index(0, length)
-        };
-
-        debug_assert_eq!(series.len(), length);
-
-        series
-    }
-
-    /// Materialize the [`ScalarColumn`] into a [`Series`].
-    pub fn to_series(&self) -> Series {
-        Self::_to_series(self.name.clone(), self.scalar.clone(), self.length)
-    }
-
-    /// Get the [`ScalarColumn`] as [`Series`]
-    ///
-    /// This needs to materialize upon the first call. Afterwards, this is cached.
-    pub fn as_materialized_series(&self) -> &Series {
-        self.materialized.get_or_init(|| self.to_series())
-    }
-
-    /// Take the [`ScalarColumn`] and materialize as a [`Series`] if not already done.
-    pub fn take_materialized_series(self) -> Series {
-        self.materialized
-            .into_inner()
-            .unwrap_or_else(|| Self::_to_series(self.name, self.scalar, self.length))
-    }
-
-    /// Take the [`ScalarColumn`] as a series with a single value.
-    ///
-    /// If the [`ScalarColumn`] has `length=0` the resulting `Series` will also have `length=0`.
-    pub fn as_single_value_series(&self) -> Series {
-        match self.materialized.get() {
-            Some(s) => s.head(Some(1)),
-            None => Self::_to_series(
-                self.name.clone(),
-                self.scalar.clone(),
-                usize::min(1, self.length),
-            ),
-        }
-    }
-
-    /// Create a new [`ScalarColumn`] from a `length=1` Series and expand it `length`.
-    ///
-    /// This will panic if the value cannot be made static or if the series has length `0`.
-    pub fn from_single_value_series(series: Series, length: usize) -> PolarsResult<Self> {
-        debug_assert_eq!(series.len(), 1);
-        let value = series.get(0)?;
-        let value = value.into_static();
-        let value = Scalar::new(series.dtype().clone(), value);
-        Ok(ScalarColumn::new(series.name().clone(), value, length))
-    }
-
-    /// Resize the [`ScalarColumn`] to new `length`.
-    ///
-    /// This reuses the materialized [`Series`], if `length <= self.length`.
-    pub fn resize(&self, length: usize) -> ScalarColumn {
-        if self.length == length {
-            return self.clone();
-        }
-
-        // This is violates an invariant if this triggers, the scalar value is undefined if the
-        // self.length == 0 so therefore we should never resize using that value.
-        debug_assert_ne!(self.length, 0);
-
-        let mut resized = Self {
-            name: self.name.clone(),
-            scalar: self.scalar.clone(),
-            length,
-            materialized: OnceLock::new(),
-        };
-
-        if self.length >= length {
-            if let Some(materialized) = self.materialized.get() {
-                resized.materialized = OnceLock::from(materialized.head(Some(length)));
-                debug_assert_eq!(resized.materialized.get().unwrap().len(), length);
-            }
-        }
-
-        resized
-    }
-
-    pub fn cast_with_options(&self, dtype: &DataType, options: CastOptions) -> PolarsResult<Self> {
-        // @NOTE: We expect that when casting the materialized series mostly does not need change
-        // the physical array. Therefore, we try to cast the entire materialized array if it is
-        // available.
-
-        match self.materialized.get() {
-            Some(s) => {
-                let materialized = s.cast_with_options(dtype, options)?;
-                assert_eq!(self.length, materialized.len());
-
-                let mut casted = if materialized.len() == 0 {
-                    Self::new_empty(materialized.name().clone(), materialized.dtype().clone())
-                } else {
-                    // SAFETY: Just did bounds check
-                    let scalar = unsafe { materialized.get_unchecked(0) }.into_static();
-                    Self::new(
-                        materialized.name().clone(),
-                        Scalar::new(materialized.dtype().clone(), scalar),
-                        self.length,
-                    )
-                };
-                casted.materialized = OnceLock::from(materialized);
-                Ok(casted)
-            },
-            None => {
-                let s = self
-                    .as_single_value_series()
-                    .cast_with_options(dtype, options)?;
-                assert_eq!(1, s.len());
-
-                if self.length == 0 {
-                    Ok(Self::new_empty(s.name().clone(), s.dtype().clone()))
-                } else {
-                    Self::from_single_value_series(s, self.length)
-                }
-            },
-        }
-    }
-
-    pub fn strict_cast(&self, dtype: &DataType) -> PolarsResult<Self> {
-        self.cast_with_options(dtype, CastOptions::Strict)
-    }
-    pub fn cast(&self, dtype: &DataType) -> PolarsResult<Self> {
-        self.cast_with_options(dtype, CastOptions::NonStrict)
-    }
-    /// # Safety
-    ///
-    /// This can lead to invalid memory access in downstream code.
-    pub unsafe fn cast_unchecked(&self, dtype: &DataType) -> PolarsResult<Self> {
-        // @NOTE: We expect that when casting the materialized series mostly does not need change
-        // the physical array. Therefore, we try to cast the entire materialized array if it is
-        // available.
-
-        match self.materialized.get() {
-            Some(s) => {
-                let materialized = s.cast_unchecked(dtype)?;
-                assert_eq!(self.length, materialized.len());
-
-                let mut casted = if materialized.len() == 0 {
-                    Self::new_empty(materialized.name().clone(), materialized.dtype().clone())
-                } else {
-                    // SAFETY: Just did bounds check
-                    let scalar = unsafe { materialized.get_unchecked(0) }.into_static();
-                    Self::new(
-                        materialized.name().clone(),
-                        Scalar::new(materialized.dtype().clone(), scalar),
-                        self.length,
-                    )
-                };
-                casted.materialized = OnceLock::from(materialized);
-                Ok(casted)
-            },
-            None => {
-                let s = self.as_single_value_series().cast_unchecked(dtype)?;
-                assert_eq!(1, s.len());
-
-                if self.length == 0 {
-                    Ok(Self::new_empty(s.name().clone(), s.dtype().clone()))
-                } else {
-                    Self::from_single_value_series(s, self.length)
-                }
-            },
-        }
-    }
-
-    pub fn has_nulls(&self) -> bool {
-        self.length != 0 && self.scalar.is_null()
     }
 }
 
@@ -1255,13 +1015,6 @@ impl IntoColumn for Column {
     #[inline(always)]
     fn into_column(self) -> Column {
         self
-    }
-}
-
-impl IntoColumn for ScalarColumn {
-    #[inline(always)]
-    fn into_column(self) -> Column {
-        self.into()
     }
 }
