+++ crates/polars-io/src/utils/other.rs
@@ -87,7 +87,7 @@ pub(crate) fn update_row_counts(dfs: &mut [(DataFrame, IdxSize)], offset: IdxSiz
         let mut previous = dfs[0].1 + offset;
         for (df, n_read) in &mut dfs[1..] {
             if let Some(s) = unsafe { df.get_columns_mut() }.get_mut(0) {
-                *s = (&*s + previous).unwrap();
+                *s = &*s + previous;
             }
             previous += *n_read;
         }
@@ -103,7 +103,7 @@ pub(crate) fn update_row_counts2(dfs: &mut [DataFrame], offset: IdxSize) {
         for df in &mut dfs[1..] {
             let n_read = df.height() as IdxSize;
             if let Some(s) = unsafe { df.get_columns_mut() }.get_mut(0) {
-                *s = (&*s + previous).unwrap();
+                *s = &*s + previous;
             }
             previous += n_read;
         }
@@ -122,7 +122,7 @@ pub(crate) fn update_row_counts3(dfs: &mut [DataFrame], heights: &[IdxSize], off
             let n_read = heights[i];
 
             if let Some(s) = unsafe { df.get_columns_mut() }.get_mut(0) {
-                *s = (&*s + previous).unwrap();
+                *s = &*s + previous;
             }
 
             previous += n_read;
