+++ crates/polars-ops/src/series/ops/duration.rs
@@ -35,7 +35,7 @@ pub fn impl_duration(s: &[Column], time_unit: TimeUnit) -> PolarsResult<Column> 
                 microseconds = (microseconds + (nanoseconds.wrapping_trunc_div_scalar(1_000)))?;
             }
             if !is_zero_scalar(&milliseconds) {
-                microseconds = (microseconds + (milliseconds * 1_000)?)?;
+                microseconds = (microseconds + milliseconds * 1_000)?;
             }
             microseconds
         },
@@ -44,10 +44,10 @@ pub fn impl_duration(s: &[Column], time_unit: TimeUnit) -> PolarsResult<Column> 
                 nanoseconds = nanoseconds.new_from_index(0, max_len);
             }
             if !is_zero_scalar(&microseconds) {
-                nanoseconds = (nanoseconds + (microseconds * 1_000)?)?;
+                nanoseconds = (nanoseconds + microseconds * 1_000)?;
             }
             if !is_zero_scalar(&milliseconds) {
-                nanoseconds = (nanoseconds + (milliseconds * 1_000_000)?)?;
+                nanoseconds = (nanoseconds + milliseconds * 1_000_000)?;
             }
             nanoseconds
         },
@@ -72,24 +72,19 @@ pub fn impl_duration(s: &[Column], time_unit: TimeUnit) -> PolarsResult<Column> 
         TimeUnit::Milliseconds => MILLISECONDS,
     };
     if !is_zero_scalar(&seconds) {
-        let units = seconds * multiplier;
-        duration = (duration + units?)?;
+        duration = (duration + seconds * multiplier)?;
     }
     if !is_zero_scalar(&minutes) {
-        let units = minutes * (multiplier * 60);
-        duration = (duration + units?)?;
+        duration = (duration + minutes * multiplier * 60)?;
     }
     if !is_zero_scalar(&hours) {
-        let units = hours * (multiplier * 60 * 60);
-        duration = (duration + units?)?;
+        duration = (duration + hours * multiplier * 60 * 60)?;
     }
     if !is_zero_scalar(&days) {
-        let units = days * (multiplier * SECONDS_IN_DAY);
-        duration = (duration + units?)?;
+        duration = (duration + days * multiplier * SECONDS_IN_DAY)?;
     }
     if !is_zero_scalar(&weeks) {
-        let units = weeks * (multiplier * SECONDS_IN_DAY * 7);
-        duration = (duration + units?)?;
+        duration = (duration + weeks * multiplier * SECONDS_IN_DAY * 7)?;
     }
 
     duration
