+++ py-polars/tests/unit/test_scalar.py
@@ -13,3 +13,26 @@ def test_invalid_broadcast() -> None:
     )
     with pytest.raises(pl.exceptions.InvalidOperationError):
         df.select(pl.col("group").filter(pl.col("group") == 0), "a")
+
+
+@pytest.mark.parametrize(
+    "dtype",
+    [
+        pl.Null,
+        pl.Int32,
+        pl.String,
+        pl.Enum(["foo"]),
+        pl.Binary,
+        pl.List(pl.Int32),
+        pl.Struct({"a": pl.Int32}),
+        pl.Array(pl.Int32, 1),
+        pl.List(pl.List(pl.Int32)),
+    ],
+)
+def test_null_literals(dtype: pl.DataType) -> None:
+    assert (
+        pl.DataFrame([pl.Series("a", [1, 2], pl.Int64)])
+        .with_columns(pl.lit(None).cast(dtype).alias("b"))
+        .collect_schema()
+        .dtypes()
+    ) == [pl.Int64, dtype]
